Implement M39.10 Rust advanced data contracts
