+++ internal_docs/typescript_go_architecture_transfer_guardrails.md
@@ -65,7 +65,7 @@ without treating probes as source content reads.
 | CLI lint command reads | `crates/sifr/src/lint_cli.rs:308`, `crates/sifr/src/lint_cli.rs:496`, `crates/sifr/src/lint_cli.rs:499` | CLI lint command reads individual files for linting and probes path exclusion/start-dir shape. | Provider-backed for semantic source reads; CLI target filtering remains a documented command-surface probe until the source-provider workstream classifies it. |
 | CLI check/package command reads | `crates/sifr/src/check_and_package_commands.rs:409`, `crates/sifr/src/check_and_package_commands.rs:415`, `crates/sifr/src/check_and_package_commands.rs:427`, `crates/sifr/src/check_and_package_commands.rs:551`, `crates/sifr/src/check_and_package_commands.rs:554`, `crates/sifr/src/check_and_package_commands.rs:587`, `crates/sifr/src/check_and_package_commands.rs:583`, `crates/sifr/src/check_and_package_commands.rs:590`, `crates/sifr/src/check_and_package_commands.rs:601`, `crates/sifr/src/check_and_package_commands.rs:658`, `crates/sifr/src/check_and_package_commands.rs:663` | CLI package/check command surfaces probe targets and cache paths and read package sources for command output. | Provider-backed for semantic source reads; command-output/cache probes remain documented exceptions where non-semantic. |
 | Package Python environment probe | `crates/sifr_package/src/cargo/python_probe.rs:11`, `crates/sifr_package/src/python/environment.rs:235`, `crates/sifr_package/src/python/environment.rs:246`, `crates/sifr_package/src/python/environment.rs:378` | Python interop checks the selected interpreter and configured `pyproject.toml`/`uv.lock` files before running the CPython probe. | Package/environment metadata probes are build identity inputs and should move behind a package-aware snapshot boundary when Python runtime metadata is integrated into broader compiler-service state. |
-| Driver Rust interop bridge/probe metadata | `crates/sifr_driver/src/build/rust_interop.rs:788`, `crates/sifr_driver/src/build/rust_interop.rs:792`, `crates/sifr_driver/src/build/rust_interop.rs:800`, `crates/sifr_driver/src/build/rust_interop/panic_validation.rs:173`, `crates/sifr_driver/src/build/rust_interop/panic_validation.rs:175`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:73`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:78`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:85`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:226`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:236`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:274`, `crates/sifr_driver/src/build/rust_interop_digest.rs:43`, `crates/sifr_driver/src/build/rust_interop_digest.rs:49`, `crates/sifr_driver/src/build/rust_interop_probe.rs:41` | Rust interop scans declared bridge roots for unsafe bridge files and shared-bridge generated-type imports, hashes bridge/Cargo cache inputs, reads Cargo profile settings for panic contracts, and probes resolved backend Cargo manifests before final build. | Package Rust interop build identity and trust-evidence probes; keep inventoried until a package-aware build snapshot owns bridge source, Cargo profile, panic profile, shared-bridge boundary, and probe inputs. |
+| Driver Rust interop bridge/probe metadata | `crates/sifr_driver/src/build/rust_interop.rs:795`, `crates/sifr_driver/src/build/rust_interop.rs:799`, `crates/sifr_driver/src/build/rust_interop.rs:807`, `crates/sifr_driver/src/build/rust_interop/panic_validation.rs:173`, `crates/sifr_driver/src/build/rust_interop/panic_validation.rs:175`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:73`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:78`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:85`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:226`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:236`, `crates/sifr_driver/src/build/rust_interop_cargo_inputs.rs:274`, `crates/sifr_driver/src/build/rust_interop_digest.rs:43`, `crates/sifr_driver/src/build/rust_interop_digest.rs:49`, `crates/sifr_driver/src/build/rust_interop_probe.rs:41` | Rust interop scans declared bridge roots for unsafe bridge files and shared-bridge generated-type imports, hashes bridge/Cargo cache inputs, reads Cargo profile settings for panic contracts, and probes resolved backend Cargo manifests before final build. | Package Rust interop build identity and trust-evidence probes; keep inventoried until a package-aware build snapshot owns bridge source, Cargo profile, panic profile, shared-bridge boundary, and probe inputs. |
 | CLI entrypoint probing | `crates/sifr/src/cli_model_and_entrypoint.rs:634`, `crates/sifr/src/cli_model_and_entrypoint.rs:690`, `crates/sifr/src/cli_model_and_entrypoint.rs:716`, `crates/sifr/src/cli_model_and_entrypoint.rs:721` | CLI mode resolution reads manifests/source files and probes sibling modules. | Provider-backed for semantic source reads; CLI mode-selection probes remain tracked lookup dependencies. |
 | CLI self-update receipt state | `crates/sifr/src/self_update_receipt.rs:66`, `crates/sifr/src/self_update_receipt.rs:85`, `crates/sifr/src/self_update_receipt.rs:96`, `crates/sifr/src/self_update_receipt.rs:105`, `crates/sifr/src/self_update_receipt.rs:160` | Self-update reads and probes the standalone installer receipt to decide whether the current executable is managed by the official installer. | Non-semantic install-state command surface; not compiler source identity. Keep inventoried unless a later install-state provider boundary is introduced. |
 | CLI self-update runner fixture reads | `crates/sifr/src/self_update_runner.rs:492`, `crates/sifr/src/self_update_runner.rs:524`, `crates/sifr/src/self_update_runner.rs:594` | Self-update runner tests read fixture files written by fake installer scripts to assert delegated environment, manifest, and lock behavior. | Non-semantic test-fixture reads; not compiler source identity. Keep inventoried unless runner tests move to a dedicated fixture helper boundary. |
