+++ verification/areas/rust_interop/data/rust_interop_fixture_matrix.json
@@ -237,9 +237,9 @@
       "tier": 4,
       "capability": "Arrow record batch exchange",
       "required_crates": ["arrow"],
-      "execution_kind": "cargo-probe",
-      "positive_evidence": {"id": "arrow_schema_identity", "status": "planned"},
-      "negative_evidence": {"id": "invalid_arrow_metadata", "status": "planned"}
+      "execution_kind": "contract-only",
+      "positive_evidence": {"id": "arrow_schema_identity", "status": "passing", "scope": "contract-only"},
+      "negative_evidence": {"id": "invalid_arrow_metadata", "status": "passing", "scope": "contract-only"}
     },
     {
       "id": "tensor_dlpack_bridge",
@@ -247,8 +247,8 @@
       "capability": "tensor and DLPack handoff",
       "required_crates": ["ndarray"],
       "execution_kind": "contract-only",
-      "positive_evidence": {"id": "explicit_tensor_ownership", "status": "planned"},
-      "negative_evidence": {"id": "implicit_dlpack_ownership_rejected", "status": "planned"}
+      "positive_evidence": {"id": "explicit_tensor_ownership", "status": "passing", "scope": "contract-only"},
+      "negative_evidence": {"id": "implicit_dlpack_ownership_rejected", "status": "passing", "scope": "contract-only"}
     },
     {
       "id": "advanced_data_matrix",
@@ -256,9 +256,9 @@
       "capability": "advanced data and tensor ecosystems",
       "required_crates": ["datafusion", "polars", "ndarray", "candle"],
       "features": {"candle": {"backend": "cpu-only"}},
-      "execution_kind": "cargo-probe",
-      "positive_evidence": {"id": "advanced_data_metadata", "status": "planned"},
-      "negative_evidence": {"id": "dtype_shape_mismatch", "status": "planned"}
+      "execution_kind": "contract-only",
+      "positive_evidence": {"id": "advanced_data_metadata", "status": "passing", "scope": "contract-only"},
+      "negative_evidence": {"id": "dtype_shape_mismatch", "status": "passing", "scope": "contract-only"}
     },
     {
       "id": "ecosystem_backend_certification",
